requirements.txt
Flask
python-dotenv
openai
linkedin-scraper
selenium
webdriver-manager